requirements.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
attrs==25.3.0
basicsr @ git+https://github.com/xinntao/BasicSR.git@8d56e3a045f9fb3e1d8872f92ee4a4f07f886b0a
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
controlnet_aux==0.0.10
cycler==0.12.1
Cython==3.1.2
diffusers==0.33.1
dill==0.4.0
easydict==1.13
einops==0.8.1
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
Flask==3.1.1
flask-cors==6.0.0
Flask-WTF==1.2.2
SQLAlchemy==2.0.30
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2024.6.1
future==1.0.0
gfpgan @ git+https://github.com/TencentARC/GFPGAN.git@7552a7791caad982045a7bbe5634bbf1cd5c8679
grpcio==1.73.0
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
insightface==0.7
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.3
numba==0.61.2
numpy==1.26.4
onnx==1.18.0
onnxruntime-gpu==1.22.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
prettytable==3.16.0
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyMatting==1.1.14
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
realesrgan==0.3.0
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.4
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
segment-anything==1.0
sentencepiece==0.2.0
simsimd==6.4.9
six==1.17.0
sounddevice==0.5.2
stringzilla==3.12.5
sympy==1.13.1
tb-nightly==2.20.0a20250609
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.6.1
timm==1.0.15
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.155
ultralytics-thop==2.0.14
urllib3==2.4.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
yapf==0.43.0
zipp==3.23.0
<<<<<<< codex/integrate-task-queue-and-progress-feedback
celery==5.3.6
redis==5.0.5
=======
SQLAlchemy 
>>>>>>> ip-adapter
